airflow_news_extraction/requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-sqlite==3.4.0
apache-airflow-providers-email==1.0.1
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
langchain==0.0.335 